+++ components/homme/src/share/cxx/GllFvRemapImpl.hpp
@@ -36,8 +36,8 @@ struct GllFvRemapImpl {
   enum : int { np2 = NP*NP };
   enum : int { num_lev_pack = NUM_LEV };
   enum : int { max_num_lev_pack = NUM_LEV_P };
-  enum : int { num_lev_aligned = num_lev_pack*packn };
-  enum : int { num_levp_aligned = max_num_lev_pack*packn };
+  enum : int { num_lev_aligned = (int)num_lev_pack*(int)packn };
+  enum : int { num_levp_aligned = (int)max_num_lev_pack*(int)packn };
   enum : int { num_phys_lev = NUM_PHYSICAL_LEV };
 
   typedef GllFvRemap::Phys1T Phys1T;
